feat: optimize project structure
