--- src-tauri/src/adapter/mod.rs
@@ -1,2 +1,0 @@
-pub mod user_adapter;
-
