+++ src-tauri/src/dao/mod.rs
@@ -1,0 +1,4 @@
+pub mod file_dao;
+pub mod user_dao;
+pub mod workspace_dao;
+
