+++ src-tauri/src/dao/workspace_dao.rs
@@ -1,4 +1,4 @@
-// src/service.rs
+// src/dao.rs
 use sea_orm::{
     ActiveModelTrait, ColumnTrait, DatabaseConnection, EntityTrait, IntoActiveModel, ModelTrait,
     QueryFilter,
