+++ src-tauri/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod api;
 pub mod entity;
+pub mod dao;
 pub mod service;
-pub mod adapter;
 pub mod util;
 
 use std::fs;
