+++ src-tauri/src/route.rs
@@ -1,9 +1,10 @@
 use base64::Engine;
 use base64::prelude::BASE64_STANDARD;
+use sea_orm::DatabaseConnection;
 use serde_json::{to_value, Value};
 use tauri::State;
 use app::{AppResponse, AppState};
-use app::adapter::user_adapter::{get_access_codes, get_user_info, login, LoginBody, logout, refresh_token, register, RegisterBody};
+use app::service::user_service::{get_access_codes, get_user_info, login, LoginBody, logout, refresh_token, register, RegisterBody};
 
 #[tauri::command]
 pub async fn route_cmd(
@@ -17,15 +18,33 @@ pub async fn route_cmd(
     }
     // Pre-processing or logging logic
     let db = &state.conn;
+    return if command.starts_with("user") {
+        Ok(invoke_user_cmd(&db, command, access_token, args).await)
+    } else if command.starts_with("workspace"){
+        // todo
+        Ok(Value::Null)
+    } else if command.starts_with("model") {
+        // todo
+        Ok(Value::Null)
+    } else {
+        let response = AppResponse::error(None::<String>, &format!("Command {:?} not found", command));
+        Ok(to_value(&response).unwrap())
+    }
+}
+
+pub async fn invoke_user_cmd(
+    db: &DatabaseConnection,
+    command: String,
+    access_token: Option<String>,
+    args: Value,
+) -> Value {
     if command == "user_login" {
         let result: LoginBody = serde_json::from_value(args).unwrap();
         let response = login(db, &result).await;
-        return Ok(to_value(&response).unwrap())
+        return to_value(&response).unwrap()
     }
     if access_token.is_none() {
-        return Ok(
-            to_value(&AppResponse::error(None::<String>, "User token is null")).unwrap(),
-        );
+        return to_value(&AppResponse::error(None::<String>, "User token is null")).unwrap()
     }
     let access_token = access_token.unwrap();
     let result = BASE64_STANDARD.decode(&access_token).unwrap();
@@ -33,30 +52,30 @@ pub async fn route_cmd(
     return match command.as_str() {
         "user_get_info" => {
             let response = get_user_info(db, &user_id).await;
-            Ok(to_value(&response).unwrap())
+            to_value(&response).unwrap()
         }
         "user_login" => {
             let result: LoginBody = serde_json::from_value(args).unwrap();
             let response = login(db, &result).await;
-            Ok(to_value(&response).unwrap())
+            to_value(&response).unwrap()
         }
         "user_logout" => {
             let response = logout().await;
-            Ok(to_value(&response).unwrap())
+            to_value(&response).unwrap()
         }
         "user_register" => {
             let result: RegisterBody = serde_json::from_value(args).unwrap();
             let response = register(db, &result).await;
-            Ok(to_value(&response).unwrap())
+            to_value(&response).unwrap()
         }
         "user_refresh_token" => {
             let response = refresh_token(db, &access_token).await;
-            Ok(to_value(&response).unwrap())
+            to_value(&response).unwrap()
         }
         "user_get_access_codes" => {
             let response = get_access_codes(db).await;
-            Ok(to_value(&response).unwrap())
+            to_value(&response).unwrap()
         }
-        _ => Ok(to_value(&AppResponse::error(None::<String>, "Command not found")).unwrap()),
+        _ => to_value(&AppResponse::error(None::<String>, "Command not found")).unwrap(),
     };
 }
