+++ src-tauri/src/service/mod.rs
@@ -1,4 +1,2 @@
-pub mod file_service;
 pub mod user_service;
-pub mod workspace_service;
